+++ lib/collection/src/events.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use segment::json_path::JsonPathV2;
+use segment::json_path::JsonPath;
 use segment::types::{Filter, PayloadFieldSchema};
 
 use crate::shards::CollectionId;
@@ -12,10 +12,10 @@ pub struct CollectionDeletedEvent {
 pub struct SlowQueryEvent {
     pub collection_id: CollectionId,
     pub filters: Vec<Filter>,
-    pub schema: HashMap<JsonPathV2, PayloadFieldSchema>,
+    pub schema: HashMap<JsonPath, PayloadFieldSchema>,
 }
 
 pub struct IndexCreatedEvent {
     pub collection_id: CollectionId,
-    pub field_name: JsonPathV2,
+    pub field_name: JsonPath,
 }
