+++ lib/segment/src/index/field_index/binary_index.rs
@@ -355,7 +355,7 @@ mod tests {
     use super::BinaryIndex;
     use crate::common::rocksdb_wrapper::open_db_with_existing_cf;
     use crate::index::field_index::{PayloadFieldIndex, ValueIndexer};
-    use crate::json_path::path;
+    use crate::json_path::JsonPath;
 
     const FIELD_NAME: &str = "bool_field";
     const DB_NAME: &str = "test_db";
@@ -370,7 +370,7 @@ mod tests {
 
     fn match_bool(value: bool) -> crate::types::FieldCondition {
         crate::types::FieldCondition::new_match(
-            path(FIELD_NAME),
+            JsonPath::new(FIELD_NAME),
             crate::types::Match::Value(crate::types::MatchValue {
                 value: crate::types::ValueVariants::Bool(value),
             }),
@@ -500,7 +500,9 @@ mod tests {
                 index.add_point(i as u32, &[&value]).unwrap();
             });
 
-        let blocks = index.payload_blocks(0, path(FIELD_NAME)).collect_vec();
+        let blocks = index
+            .payload_blocks(0, JsonPath::new(FIELD_NAME))
+            .collect_vec();
         assert_eq!(blocks.len(), 2);
         assert_eq!(blocks[0].cardinality, 6);
         assert_eq!(blocks[1].cardinality, 6);
