+++ lib/segment/src/index/field_index/full_text_index/text_index.rs
@@ -243,10 +243,10 @@ mod tests {
     use super::*;
     use crate::common::rocksdb_wrapper::open_db_with_existing_cf;
     use crate::data_types::index::{TextIndexType, TokenizerType};
-    use crate::json_path::path;
+    use crate::json_path::JsonPath;
 
     fn filter_request(text: &str) -> FieldCondition {
-        FieldCondition::new_match(path("text"), Match::new_text(text))
+        FieldCondition::new_match(JsonPath::new("text"), Match::new_text(text))
     }
 
     #[rstest]
