+++ lib/segment/src/index/plain_payload_index.rs
@@ -86,8 +86,10 @@ impl PayloadIndex for PlainPayloadIndex {
     fn set_indexed(
         &mut self,
         field: PayloadKeyTypeRef,
-        payload_schema: PayloadFieldSchema,
+        payload_schema: impl Into<PayloadFieldSchema>,
     ) -> OperationResult<()> {
+        let payload_schema = payload_schema.into();
+
         if let Some(prev_schema) = self
             .config
             .indexed_fields
