+++ lib/segment/src/payload_storage/payload_storage_enum.rs
@@ -130,7 +130,6 @@ mod tests {
 
     use super::*;
     use crate::common::rocksdb_wrapper::{open_db, DB_VECTOR_CF};
-    use crate::json_path::path;
     use crate::types::Payload;
 
     #[test]
@@ -178,7 +177,7 @@ mod tests {
             let partial_payload: Payload = serde_json::from_str(r#"{ "age": 53 }"#).unwrap();
             storage.assign(100, &partial_payload).unwrap();
 
-            storage.delete(100, &path("location.geo")).unwrap();
+            storage.delete(100, &JsonPath::new("location.geo")).unwrap();
 
             let res = storage.payload(100).unwrap();
 
@@ -202,8 +201,10 @@ mod tests {
                 serde_json::from_str(r#"{ "hobby": "vector search" }"#).unwrap();
             storage.assign(100, &partial_payload).unwrap();
 
-            storage.delete(100, &path("location.city")).unwrap();
-            storage.delete(100, &path("location")).unwrap();
+            storage
+                .delete(100, &JsonPath::new("location.city"))
+                .unwrap();
+            storage.delete(100, &JsonPath::new("location")).unwrap();
 
             let res = storage.payload(100).unwrap();
