+++ lib/segment/src/payload_storage/query_checker.rs
@@ -11,7 +11,6 @@ use common::types::PointOffsetType;
 use crate::common::utils::{check_is_empty, check_is_null, IndexesMap};
 use crate::id_tracker::IdTrackerSS;
 use crate::index::field_index::FieldIndex;
-use crate::json_path::JsonPathInterface as _;
 use crate::payload_storage::condition_checker::ValueChecker;
 use crate::payload_storage::payload_storage_enum::PayloadStorageEnum;
 use crate::payload_storage::ConditionChecker;
@@ -296,7 +295,7 @@ mod tests {
     use crate::common::rocksdb_wrapper::{open_db, DB_VECTOR_CF};
     use crate::id_tracker::simple_id_tracker::SimpleIdTracker;
     use crate::id_tracker::IdTracker;
-    use crate::json_path::path;
+    use crate::json_path::JsonPath;
     use crate::payload_storage::simple_payload_storage::SimplePayloadStorage;
     use crate::payload_storage::PayloadStorage;
     use crate::types::{
@@ -342,72 +341,78 @@ mod tests {
         );
 
         let is_empty_condition = Filter::new_must(Condition::IsEmpty(IsEmptyCondition {
-            is_empty: PayloadField { key: path("price") },
+            is_empty: PayloadField {
+                key: JsonPath::new("price"),
+            },
         }));
         assert!(!payload_checker.check(0, &is_empty_condition));
 
         let is_empty_condition = Filter::new_must(Condition::IsEmpty(IsEmptyCondition {
             is_empty: PayloadField {
-                key: path("something_new"),
+                key: JsonPath::new("something_new"),
             },
         }));
         assert!(payload_checker.check(0, &is_empty_condition));
 
         let is_empty_condition = Filter::new_must(Condition::IsEmpty(IsEmptyCondition {
-            is_empty: PayloadField { key: path("parts") },
+            is_empty: PayloadField {
+                key: JsonPath::new("parts"),
+            },
         }));
         assert!(payload_checker.check(0, &is_empty_condition));
 
         let is_empty_condition = Filter::new_must(Condition::IsEmpty(IsEmptyCondition {
             is_empty: PayloadField {
-                key: path("not_null"),
+                key: JsonPath::new("not_null"),
             },
         }));
         assert!(!payload_checker.check(0, &is_empty_condition));
 
         let is_null_condition = Filter::new_must(Condition::IsNull(IsNullCondition {
             is_null: PayloadField {
-                key: path("amount"),
-            },
-        }));
-        assert!(!payload_checker.check(0, &is_null_condition));
-
-        let is_null_condition = Filter::new_must(Condition::IsNull(IsNullCondition {
-            is_null: PayloadField { key: path("parts") },
-        }));
-        assert!(!payload_checker.check(0, &is_null_condition));
-
-        let is_null_condition = Filter::new_must(Condition::IsNull(IsNullCondition {
-            is_null: PayloadField {
-                key: path("something_else"),
+                key: JsonPath::new("amount"),
             },
         }));
         assert!(!payload_checker.check(0, &is_null_condition));
 
         let is_null_condition = Filter::new_must(Condition::IsNull(IsNullCondition {
             is_null: PayloadField {
-                key: path("packaging"),
+                key: JsonPath::new("parts"),
+            },
+        }));
+        assert!(!payload_checker.check(0, &is_null_condition));
+
+        let is_null_condition = Filter::new_must(Condition::IsNull(IsNullCondition {
+            is_null: PayloadField {
+                key: JsonPath::new("something_else"),
+            },
+        }));
+        assert!(!payload_checker.check(0, &is_null_condition));
+
+        let is_null_condition = Filter::new_must(Condition::IsNull(IsNullCondition {
+            is_null: PayloadField {
+                key: JsonPath::new("packaging"),
             },
         }));
         assert!(payload_checker.check(0, &is_null_condition));
 
         let is_null_condition = Filter::new_must(Condition::IsNull(IsNullCondition {
             is_null: PayloadField {
-                key: path("not_null"),
+                key: JsonPath::new("not_null"),
             },
         }));
         assert!(!payload_checker.check(0, &is_null_condition));
 
         let match_red = Condition::Field(FieldCondition::new_match(
-            path("color"),
+            JsonPath::new("color"),
             "red".to_owned().into(),
         ));
         let match_blue = Condition::Field(FieldCondition::new_match(
-            path("color"),
+            JsonPath::new("color"),
             "blue".to_owned().into(),
         ));
         let shipped_in_february = Condition::Field(FieldCondition::new_datetime_range(
-            path("shipped_at"),
+            JsonPath::new("shipped_at"),
             Range {
                 lt: Some(DateTimeWrapper::from_str("2020-03-01T00:00:00Z").unwrap()),
                 gt: None,
@@ -416,7 +421,7 @@ mod tests {
             },
         ));
         let shipped_in_march = Condition::Field(FieldCondition::new_datetime_range(
-            path("shipped_at"),
+            JsonPath::new("shipped_at"),
             Range {
                 lt: Some(DateTimeWrapper::from_str("2020-04-01T00:00:00Z").unwrap()),
                 gt: None,
@@ -424,12 +429,14 @@ mod tests {
                 lte: None,
             },
         ));
-        let with_delivery =
-            Condition::Field(FieldCondition::new_match(path("has_delivery"), true.into()));
+        let with_delivery = Condition::Field(FieldCondition::new_match(
+            JsonPath::new("has_delivery"),
+            true.into(),
+        ));
 
         let many_value_count_condition =
             Filter::new_must(Condition::Field(FieldCondition::new_values_count(
-                path("rating"),
+                JsonPath::new("rating"),
                 ValuesCount {
                     lt: None,
                     gt: None,
@@ -441,7 +448,7 @@ mod tests {
 
         let few_value_count_condition =
             Filter::new_must(Condition::Field(FieldCondition::new_values_count(
-                path("rating"),
+                JsonPath::new("rating"),
                 ValuesCount {
                     lt: Some(5),
                     gt: None,
@@ -452,7 +459,7 @@ mod tests {
         assert!(payload_checker.check(0, &few_value_count_condition));
 
         let in_berlin = Condition::Field(FieldCondition::new_geo_bounding_box(
-            path("location"),
+            JsonPath::new("location"),
             GeoBoundingBox {
                 top_left: GeoPoint {
                     lon: 13.08835,
@@ -466,7 +473,7 @@ mod tests {
         ));
 
         let in_moscow = Condition::Field(FieldCondition::new_geo_bounding_box(
-            path("location"),
+            JsonPath::new("location"),
             GeoBoundingBox {
                 top_left: GeoPoint {
                     lon: 37.0366,
@@ -480,7 +487,7 @@ mod tests {
         ));
 
         let with_bad_rating = Condition::Field(FieldCondition::new_range(
-            path("rating"),
+            JsonPath::new("rating"),
             Range {
                 lt: None,
                 gt: None,
