+++ lib/segment/src/segment.rs
@@ -1684,7 +1684,7 @@ impl SegmentEntry for Segment {
                     segment
                         .payload_index
                         .borrow_mut()
-                        .set_indexed(key, schema_type.into())?;
+                        .set_indexed(key, schema_type)?;
                     Ok(true)
                 }
             },
