+++ lib/segment/tests/integration/multivector_hnsw_test.rs
@@ -17,7 +17,7 @@ use segment::fixtures::payload_fixtures::random_int_payload;
 use segment::index::hnsw_index::graph_links::GraphLinksRam;
 use segment::index::hnsw_index::hnsw::{HNSWIndex, HnswIndexOpenArgs};
 use segment::index::VectorIndex;
-use segment::json_path::path;
+use segment::json_path::JsonPath;
 use segment::segment_constructor::build_segment;
 use segment::spaces::metric::Metric;
 use segment::spaces::simple::{CosineMetric, DotProductMetric, EuclidMetric, ManhattanMetric};
@@ -63,7 +63,11 @@ fn test_single_multi_and_dense_hnsw_equivalency() {
     let mut segment = build_segment(dir.path(), &config, true).unwrap();
 
     segment
-        .create_field_index(0, &path(int_key), Some(&PayloadSchemaType::Integer.into()))
+        .create_field_index(
+            0,
+            &JsonPath::new(int_key),
+            Some(&PayloadSchemaType::Integer.into()),
+        )
         .unwrap();
 
     let dir = Builder::new().prefix("storage_dir").tempdir().unwrap();
@@ -172,7 +176,7 @@ fn test_single_multi_and_dense_hnsw_equivalency() {
         let payload_value = random_int_payload(&mut rnd, 1..=1).pop().unwrap();
 
         let filter = Filter::new_must(Condition::Field(FieldCondition::new_match(
-            path(int_key),
+            JsonPath::new(int_key),
             payload_value.into(),
         )));
