+++ lib/segment/tests/integration/nested_filtering_test.rs
@@ -6,7 +6,7 @@ use common::types::PointOffsetType;
 use segment::fixtures::payload_context_fixture::FixtureIdTracker;
 use segment::index::struct_payload_index::StructPayloadIndex;
 use segment::index::PayloadIndex;
-use segment::json_path::path;
+use segment::json_path::JsonPath;
 use segment::payload_storage::in_memory_payload_storage::InMemoryPayloadStorage;
 use segment::payload_storage::PayloadStorage;
 use segment::types::{Condition, FieldCondition, Filter, Match, Payload, PayloadSchemaType, Range};
@@ -75,37 +75,37 @@ fn test_filtering_context_consistency() {
         StructPayloadIndex::open(wrapped_payload_storage, id_tracker, dir.path(), true).unwrap();
 
     index
-        .set_indexed(&path("f"), PayloadSchemaType::Integer.into())
+        .set_indexed(&JsonPath::new("f"), PayloadSchemaType::Integer)
         .unwrap();
     index
-        .set_indexed(&path("arr1[].a"), PayloadSchemaType::Integer.into())
+        .set_indexed(&JsonPath::new("arr1[].a"), PayloadSchemaType::Integer)
         .unwrap();
     index
-        .set_indexed(&path("arr1[].b"), PayloadSchemaType::Integer.into())
+        .set_indexed(&JsonPath::new("arr1[].b"), PayloadSchemaType::Integer)
         .unwrap();
     index
-        .set_indexed(&path("arr1[].c"), PayloadSchemaType::Integer.into())
+        .set_indexed(&JsonPath::new("arr1[].c"), PayloadSchemaType::Integer)
         .unwrap();
     index
-        .set_indexed(&path("arr1[].d"), PayloadSchemaType::Integer.into())
+        .set_indexed(&JsonPath::new("arr1[].d"), PayloadSchemaType::Integer)
         .unwrap();
     index
-        .set_indexed(&path("arr1[].text"), PayloadSchemaType::Text.into())
+        .set_indexed(&JsonPath::new("arr1[].text"), PayloadSchemaType::Text)
         .unwrap();
 
     {
         let nested_condition_0 = Condition::new_nested(
-            path("arr1"),
+            JsonPath::new("arr1"),
             Filter {
                 must: Some(vec![
                     // E.g. idx = 6 => { "a" = 1, "b" = 7, "c" = 1, "d" = 0 }
-                    Condition::Field(FieldCondition::new_match(path("a"), 1.into())),
-                    Condition::Field(FieldCondition::new_match(path("c"), 1.into())),
+                    Condition::Field(FieldCondition::new_match(JsonPath::new("a"), 1.into())),
+                    Condition::Field(FieldCondition::new_match(JsonPath::new("c"), 1.into())),
                 ]),
                 should: None,
                 min_should: None,
                 must_not: Some(vec![Condition::Field(FieldCondition::new_range(
-                    path("d"),
+                    JsonPath::new("d"),
                     Range {
                         lte: Some(1.into()),
                         ..Default::default()
@@ -137,13 +137,13 @@ fn test_filtering_context_consistency() {
 
     {
         let nested_condition_1 = Condition::new_nested(
-            path("arr1"),
+            JsonPath::new("arr1"),
             Filter {
                 must: Some(vec![
                     // E.g. idx = 6 => { "a" = 1, "b" = 7, "c" = 1, "d" = 0 }
-                    Condition::Field(FieldCondition::new_match(path("a"), 1.into())),
-                    Condition::Field(FieldCondition::new_match(path("c"), 1.into())),
-                    Condition::Field(FieldCondition::new_match(path("d"), 0.into())),
+                    Condition::Field(FieldCondition::new_match(JsonPath::new("a"), 1.into())),
+                    Condition::Field(FieldCondition::new_match(JsonPath::new("c"), 1.into())),
+                    Condition::Field(FieldCondition::new_match(JsonPath::new("d"), 0.into())),
                 ]),
                 should: None,
                 min_should: None,
@@ -169,16 +169,16 @@ fn test_filtering_context_consistency() {
 
     {
         let nested_condition_2 = Condition::new_nested(
-            path("arr1"),
+            JsonPath::new("arr1"),
             Filter {
                 must: Some(vec![
                     // E.g. idx = 6 => { "a" = 1, "b" = 7, "c" = 1, "d" = 0 }
-                    Condition::Field(FieldCondition::new_match(path("a"), 1.into())),
+                    Condition::Field(FieldCondition::new_match(JsonPath::new("a"), 1.into())),
                     Condition::Field(FieldCondition::new_match(
-                        path("text"),
+                        JsonPath::new("text"),
                         Match::Text("c1".to_string().into()),
                     )),
-                    Condition::Field(FieldCondition::new_match(path("d"), 0.into())),
+                    Condition::Field(FieldCondition::new_match(JsonPath::new("d"), 0.into())),
                 ]),
                 should: None,
                 min_should: None,
@@ -203,20 +203,20 @@ fn test_filtering_context_consistency() {
 
     {
         let nested_condition_3 = Condition::new_nested(
-            path("arr1"),
+            JsonPath::new("arr1"),
             Filter::new_must(Condition::Field(FieldCondition::new_match(
-                path("b"),
+                JsonPath::new("b"),
                 1.into(),
             ))),
         );
 
         let nester_condition_3_1 = Condition::new_nested(
-            path("arr2"),
+            JsonPath::new("arr2"),
             Filter {
                 must: Some(vec![Condition::new_nested(
-                    path("arr3"),
+                    JsonPath::new("arr3"),
                     Filter::new_must(Condition::Field(FieldCondition::new_match(
-                        path("b"),
+                        JsonPath::new("b"),
                         10.into(),
                     ))),
                 )]),
