+++ lib/segment/tests/integration/payload_index_test.rs
@@ -21,7 +21,7 @@ use segment::fixtures::payload_fixtures::{
 use segment::index::field_index::{FieldIndex, PrimaryCondition};
 use segment::index::struct_payload_index::StructPayloadIndex;
 use segment::index::PayloadIndex;
-use segment::json_path::path;
+use segment::json_path::JsonPath;
 use segment::payload_storage::in_memory_payload_storage::InMemoryPayloadStorage;
 use segment::payload_storage::PayloadStorage;
 use segment::segment::Segment;
@@ -70,7 +70,7 @@ fn build_test_segments(path_struct: &Path, path_plain: &Path) -> (Segment, Segme
 
     let mut opnum = 0;
     struct_segment
-        .create_field_index(opnum, &path(INT_KEY_2), Some(&Integer.into()))
+        .create_field_index(opnum, &JsonPath::new(INT_KEY_2), Some(&Integer.into()))
         .unwrap();
 
     opnum += 1;
@@ -96,15 +96,15 @@ fn build_test_segments(path_struct: &Path, path_plain: &Path) -> (Segment, Segme
     }
 
     struct_segment
-        .create_field_index(opnum, &path(STR_KEY), Some(&Keyword.into()))
+        .create_field_index(opnum, &JsonPath::new(STR_KEY), Some(&Keyword.into()))
         .unwrap();
     struct_segment
-        .create_field_index(opnum, &path(INT_KEY), None)
+        .create_field_index(opnum, &JsonPath::new(INT_KEY), None)
         .unwrap();
     struct_segment
         .create_field_index(
             opnum,
-            &path(INT_KEY_2),
+            &JsonPath::new(INT_KEY_2),
             Some(&FieldParams(PayloadSchemaParams::Integer(
                 IntegerIndexParams {
                     r#type: IntegerIndexType::Integer,
@@ -117,7 +117,7 @@ fn build_test_segments(path_struct: &Path, path_plain: &Path) -> (Segment, Segme
     struct_segment
         .create_field_index(
             opnum,
-            &path(INT_KEY_3),
+            &JsonPath::new(INT_KEY_3),
             Some(&FieldParams(PayloadSchemaParams::Integer(
                 IntegerIndexParams {
                     r#type: IntegerIndexType::Integer,
@@ -128,17 +128,21 @@ fn build_test_segments(path_struct: &Path, path_plain: &Path) -> (Segment, Segme
         )
         .unwrap();
     struct_segment
-        .create_field_index(opnum, &path(GEO_KEY), Some(&PayloadSchemaType::Geo.into()))
+        .create_field_index(
+            opnum,
+            &JsonPath::new(GEO_KEY),
+            Some(&PayloadSchemaType::Geo.into()),
+        )
         .unwrap();
     struct_segment
         .create_field_index(
             opnum,
-            &path(TEXT_KEY),
+            &JsonPath::new(TEXT_KEY),
             Some(&PayloadSchemaType::Text.into()),
         )
         .unwrap();
     struct_segment
-        .create_field_index(opnum, &path(FLICKING_KEY), Some(&Integer.into()))
+        .create_field_index(opnum, &JsonPath::new(FLICKING_KEY), Some(&Integer.into()))
         .unwrap();
 
     for _ in 0..points_to_clear {
@@ -206,9 +210,10 @@ fn build_test_segments_nested_payload(path_struct: &Path, path_plain: &Path) -> 
     let points_to_clear = 500;
 
     // Nested payload keys
-    let nested_str_key = path(&format!("{}.{}.{}", STR_KEY, "nested_1", "nested_2"));
-    let nested_str_proj_key = path(&format!("{}.{}[].{}", STR_PROJ_KEY, "nested_1", "nested_2"));
-    let deep_nested_str_proj_key = path(&format!(
+    let nested_str_key = JsonPath::new(&format!("{}.{}.{}", STR_KEY, "nested_1", "nested_2"));
+    let nested_str_proj_key =
+        JsonPath::new(&format!("{}.{}[].{}", STR_PROJ_KEY, "nested_1", "nested_2"));
+    let deep_nested_str_proj_key = JsonPath::new(&format!(
         "{}[].{}[].{}",
         STR_ROOT_PROJ_KEY, "nested_1", "nested_2"
     ));
@@ -360,7 +365,7 @@ fn test_is_empty_conditions() {
 
     let filter = Filter::new_must(Condition::IsEmpty(IsEmptyCondition {
         is_empty: PayloadField {
-            key: path(FLICKING_KEY),
+            key: JsonPath::new(FLICKING_KEY),
         },
     }));
 
@@ -409,7 +414,7 @@ fn test_integer_index_types() {
     assert!(matches!(
         indexes
             .field_indexes
-            .get(&path(INT_KEY))
+            .get(&JsonPath::new(INT_KEY))
             .unwrap()
             .as_slice(),
         [FieldIndex::IntMapIndex(_), FieldIndex::IntIndex(_)]
@@ -417,7 +422,7 @@ fn test_integer_index_types() {
     assert!(matches!(
         indexes
             .field_indexes
-            .get(&path(INT_KEY_2))
+            .get(&JsonPath::new(INT_KEY_2))
             .unwrap()
             .as_slice(),
         [FieldIndex::IntMapIndex(_)]
@@ -425,7 +430,7 @@ fn test_integer_index_types() {
     assert!(matches!(
         indexes
             .field_indexes
-            .get(&path(INT_KEY_3))
+            .get(&JsonPath::new(INT_KEY_3))
             .unwrap()
             .as_slice(),
         [FieldIndex::IntIndex(_)]
@@ -440,7 +445,7 @@ fn test_cardinality_estimation() {
     let (struct_segment, _) = build_test_segments(dir1.path(), dir2.path());
 
     let filter = Filter::new_must(Condition::Field(FieldCondition::new_range(
-        path(INT_KEY),
+        JsonPath::new(INT_KEY),
         Range {
             lt: None,
             gt: None,
@@ -480,9 +485,10 @@ fn test_root_nested_array_filter_cardinality_estimation() {
 
     // rely on test data from `build_test_segments_nested_payload`
     let nested_key = "nested_1[].nested_2";
-    let nested_match = FieldCondition::new_match(path(nested_key), "some value".to_owned().into());
+    let nested_match =
+        FieldCondition::new_match(JsonPath::new(nested_key), "some value".to_owned().into());
     let filter = Filter::new_must(Condition::new_nested(
-        path(STR_ROOT_PROJ_KEY),
+        JsonPath::new(STR_ROOT_PROJ_KEY),
         Filter::new_must(Condition::Field(nested_match)),
     ));
 
@@ -497,7 +503,7 @@ fn test_root_nested_array_filter_cardinality_estimation() {
     let primary_clause = estimation.primary_clauses.first().unwrap();
 
     let expected_primary_clause = FieldCondition::new_match(
-        path(&format!("{}[].{}", STR_ROOT_PROJ_KEY, nested_key)), // full key expected
+        JsonPath::new(&format!("{}[].{}", STR_ROOT_PROJ_KEY, nested_key)), // full key expected
         "some value".to_owned().into(),
     );
 
@@ -534,12 +540,14 @@ fn test_nesting_nested_array_filter_cardinality_estimation() {
 
     // rely on test data from `build_test_segments_nested_payload`
     let nested_match_key = "nested_2";
-    let nested_match =
-        FieldCondition::new_match(path(nested_match_key), "some value".to_owned().into());
+    let nested_match = FieldCondition::new_match(
+        JsonPath::new(nested_match_key),
+        "some value".to_owned().into(),
+    );
     let filter = Filter::new_must(Condition::new_nested(
-        path(STR_ROOT_PROJ_KEY),
+        JsonPath::new(STR_ROOT_PROJ_KEY),
         Filter::new_must(Condition::new_nested(
-            path("nested_1"),
+            JsonPath::new("nested_1"),
             Filter::new_must(Condition::Field(nested_match)),
         )),
     ));
@@ -556,7 +564,7 @@ fn test_nesting_nested_array_filter_cardinality_estimation() {
 
     let expected_primary_clause = FieldCondition::new_match(
         // full key expected
-        path(&format!(
+        JsonPath::new(&format!(
             "{}[].nested_1[].{}",
             STR_ROOT_PROJ_KEY, nested_match_key
         )),
@@ -672,7 +680,7 @@ fn test_struct_payload_geo_boundingbox_index() {
     };
 
     let condition = Condition::Field(FieldCondition::new_geo_bounding_box(
-        path("geo_key"),
+        JsonPath::new("geo_key"),
         geo_bbox,
     ));
 
@@ -694,7 +702,10 @@ fn test_struct_payload_geo_radius_index() {
         radius: r_meters,
     };
 
-    let condition = Condition::Field(FieldCondition::new_geo_radius(path("geo_key"), geo_radius));
+    let condition = Condition::Field(FieldCondition::new_geo_radius(
+        JsonPath::new("geo_key"),
+        geo_radius,
+    ));
 
     let query_filter = Filter::new_must(condition);
 
@@ -733,7 +744,7 @@ fn test_struct_payload_geo_polygon_index() {
     };
 
     let condition = Condition::Field(FieldCondition::new_geo_polygon(
-        path("geo_key"),
+        JsonPath::new("geo_key"),
         geo_polygon,
     ));
 
@@ -838,10 +849,10 @@ fn test_update_payload_index_type() {
     let mut index =
         StructPayloadIndex::open(wrapped_payload_storage, id_tracker, dir.path(), true).unwrap();
 
-    let field = path("field");
+    let field = JsonPath::new("field");
 
     // set field to Integer type
-    index.set_indexed(&field, Integer.into()).unwrap();
+    index.set_indexed(&field, Integer).unwrap();
     assert_eq!(
         *index.indexed_fields().get(&field).unwrap(),
         FieldType(Integer)
@@ -851,7 +862,7 @@ fn test_update_payload_index_type() {
     assert_eq!(field_index[1].count_indexed_points(), point_num);
 
     // update field to Keyword type
-    index.set_indexed(&field, Keyword.into()).unwrap();
+    index.set_indexed(&field, Keyword).unwrap();
     assert_eq!(
         *index.indexed_fields().get(&field).unwrap(),
         FieldType(Keyword)
@@ -860,7 +871,7 @@ fn test_update_payload_index_type() {
     assert_eq!(field_index[0].count_indexed_points(), 0); // only one field index for Keyword
 
     // set field to Integer type (again)
-    index.set_indexed(&field, Integer.into()).unwrap();
+    index.set_indexed(&field, Integer).unwrap();
     assert_eq!(
         *index.indexed_fields().get(&field).unwrap(),
         FieldType(Integer)
@@ -880,7 +891,7 @@ fn test_any_matcher_cardinality_estimation() {
     let keywords: IndexSet<String, FnvBuildHasher> =
         ["value1", "value2"].iter().map(|i| i.to_string()).collect();
     let any_match = FieldCondition::new_match(
-        path(STR_KEY),
+        JsonPath::new(STR_KEY),
         Match::new_any(AnyVariants::Keywords(keywords)),
     );
